requirements.txt
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
Django==3.2
django-cors-headers==3.7.0
django-extensions==3.1.2
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
drf-extensions==0.7.0
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
pycparser==2.20
pytz==2021.1
regex==2021.4.4
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
